Fixes bug with overwrite prompt.
If a file already exists in the trashcan, it will prompt to overwrite.
A bug caused it to always overwrite as long as you didn't answer with
"n" or "N", so if you answered with "k" or "l" or whatever, it would
still overwrite. Basically, any letter that's not "n" or "N" or be
interpretted as a "y". Fixed.

Also, the whole script basically needs to be refactored.